Added Widi device and name aliases
Added aliases so we can change the 'name' field depending on where
devices are plugged into.  the main problem comes when the fighter
twister is either plugged direct into the morningstar controller vs
direct into the computer.  By changing the midi name, we can swap which
midi port it is listening to.

Added the Widi port which sends all incoming messages direct to the
Axefx.  This also will sellectively direct messages from the morningstar
to the fighter twister if the fighter twister is plugged into the
computer instead of the morningstar.  lastly, this will update volume
exp pedal and reverb exp pedal values on the fighter twister when these
get sent to the axefx for visual feedback.

Also when block channel changes or bypass toggle states are sent from
the morningstar, these also update on the fighter twister for added
visual feedback.