Added CLI args and logging on GUI
GUI now displays the last log message sent so we can see some more data
behind the scenes in the GUI window.

Added CLI arguments to change the midi ports for the devices being used.
We have 3 profiles, one for standard wireless operation (default), one
for when the MC6Pro is plugged into the computer (not using wireless) and
one where both the MC6Pro and the Fighter Twister are both plugged into
the computer.